Fixing poor package name and app compat version
