+++ digitsmigrationhelpers/build.gradle
@@ -63,7 +63,7 @@ dependencies {
     provided 'com.google.firebase:firebase-auth:11.0.0'
     compile fileTree(dir: 'libs', include: ['*.jar'])
     //noinspection GradleDependency
-    compile 'com.android.support:appcompat-v7:25.1.1'
+    compile 'com.android.support:appcompat-v7:25.2.0'
     testCompile 'junit:junit:4.12'
     testCompile 'org.robolectric:robolectric:3.0'
     testCompile "org.mockito:mockito-core:1.10.0"
