+++ digitsmigrationhelpers/src/test/java/com/firebase/digitsmigrationhelpers/RobolectricGradleTestRunner.java
@@ -11,7 +11,7 @@
  * express or implied. See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.digits.auth.migration;
+package com.firebase.digitsmigrationhelpers;
 
 import org.junit.runners.model.InitializationError;
 import org.robolectric.RobolectricTestRunner;
